+++ src/pages/PostDetail/PostCard/PostCard.jsx
@@ -1,22 +1,17 @@
 import axios from 'axios';
-import { useCallback, useRef, useState, useEffect } from 'react';
+import { useState, useEffect } from 'react';
 import { useLocation } from 'react-router-dom';
 import Post from '../../../components/card/Post';
-import { CardWrap, Form, ProfileImg, Input, BtnSubmit } from './postCardStyle';
+import { CardWrap } from './postCardStyle';
 import HeaderBM from '../../../components/header/HeaderBM';
 import CommentWrapper from '../../../components/comment/CommentWrapper';
+import CommentForm from '../../../components/comment/CommentForm';
 
 const PostCard = () => {
-  const textRef = useRef();
-  const handleResizeHeight = useCallback(() => {
-    textRef.current.style.height = '1px';
-    textRef.current.style.height = `${textRef.current.scrollHeight}px`;
-  }, []);
-
   const [posts, setPosts] = useState('');
   const location = useLocation();
   const postDetailId = { ...location.state };
-  console.log('getPostDetail 데이터 : ', posts);
+  // console.log('getPostDetail 데이터 : ', posts);
 
   const getPostDetail = async () => {
     const URL = 'https://mandarin.api.weniv.co.kr';
@@ -30,8 +25,8 @@ const PostCard = () => {
         },
       });
       setPosts(res?.data.post);
-      console.log('getPostDetail 응답 : ', res);
-      console.log(postDetailId);
+      // console.log('getPostDetail 응답 : ', res);
+      // console.log(postDetailId);
     } catch (error) {
       console.log(error);
     }
@@ -44,16 +39,7 @@ const PostCard = () => {
       <CardWrap>
         {posts && <Post posts={posts} />}
         <CommentWrapper posts={posts} postDetailId={postDetailId} />
-        <Form>
-          <ProfileImg />
-          <Input
-            placeholder="댓글 입력하기..."
-            ref={textRef}
-            onInput={handleResizeHeight}
-            rows="1"
-          />
-          <BtnSubmit type="submit">게시</BtnSubmit>
-        </Form>
+        <CommentForm />
       </CardWrap>
     </>
   );
